// src/app/shared/models/cow.interface.ts
export interface Cow {
  ageInDays: number;
  animalId: string;
  cowId: number;
  daysInLactation: number;
  deletable: boolean;
  endDate: number;
  eventId: number;
  healthIndex: number;
  lactationNumber: number;
  minValueDateTime: number;
  reportingDateTime: number;
  startDateTime: number;
  type: string;
}